Simplify the tool applet names and merge the convert and plot tools
Applets are now named for what they do, without a verb prefix or a noun
suffix: segment (was add-segmentations), calibrate (calibrate-cameras),
gpu (check-gpu), depth (compute-depth), disparity (compute-disparity),
mosaic (create-mosaic), extract (extract-frames), configs (get-configs),
view (launch-annotator) and search (launch-search). The scripts carry the
same names.

Two tools merge:
- viame convert handles stereo calibration formats (the former
  convert-cam) and ITK HDF5 transforms (convert-itk): a .h5 input, or
  -i itk, takes the transform path with its --left/--right camera names.
  cv2 and scipy are imported only on the calibration path and h5py only
  on the transform path, so neither pulls in the other's dependencies.
- viame plot has two subcommands: detections (the former plot-detections,
  now with -input/-output/-ext options) and eval (plot-eval). run_bulk's
  detection plots call the same function.

The inspect applet's script becomes inspect_file.py: a tool run from the
configs folder had Python's stdlib inspect module shadowed by it, which
broke any tool importing scipy (viame convert, mosaic). Also fix the
detection plots' savefig call for current matplotlib (bbox_extra_artists).

Project scripts, examples and READMEs use the new names; DIVE (pinned at
viame/main 0043ceef) looks for convert.py and falls back to convert_cam.py.
