+++ configs/add-ons/arctic-seal/detector_arctic_seal_yolo_2-cam.pipe
@@ -6,7 +6,7 @@
 # optical stream; thermal detections are mapped into the optical frame using
 # a camera registration file (DIVE camera registration .json or plain text
 # 3x3 homography; legacy ITK .h5 files must first be converted with
-# tools/convert_itk.py).
+# viame convert).
 #
 # Camera Order: EO, IR
 # Camera 1 is optical and camera 2 is thermal; DIVE matches the dataset's
