+++ configs/add-ons/arctic-seal/detector_arctic_seal_yolo_3-cam.pipe
@@ -6,7 +6,7 @@
 # detection on the optical stream; detections are mapped between cameras
 # using camera registration files (DIVE camera registration .json or plain
 # text 3x3 homography; legacy ITK .h5 files must first be converted with
-# tools/convert_itk.py).
+# viame convert).
 #
 # Camera Order: EO, UV, IR
 # Camera 1 is optical, camera 2 is ultraviolet, and camera 3 is thermal; DIVE
