+++ configs/prj-linux/launch_search_interface.sh
@@ -6,5 +6,5 @@ export VIAME_INSTALL=/opt/noaa/viame
 # Setup paths and run command
 source ${VIAME_INSTALL}/setup_viame.sh
 
-python ${VIAME_INSTALL}/configs/launch_search.py \
+python ${VIAME_INSTALL}/configs/search.py \
   -qp pipelines/query_retrieval_and_iqr.pipe
