+++ configs/prj-linux/launch_view_interface.sh
@@ -14,5 +14,5 @@ source ${VIAME_INSTALL}/setup_viame.sh
 # Set current directory for project folder pipe
 export VIAME_PROJECT_DIR="$(cd "$(dirname ${BASH_SOURCE[0]})" && pwd)"
 
-python ${VIAME_INSTALL}/configs/launch_annotator.py \
+python ${VIAME_INSTALL}/configs/view.py \
   -d ${INPUT_DIRECTORY} -c ${CACHE_DIRECTORY} -frate ${FRAME_RATE}
