+++ configs/prj-windows/launch_search_interface.bat
@@ -6,5 +6,5 @@ SET VIAME_INSTALL=C:\Program Files\VIAME
 REM Setup paths and run command
 CALL "%VIAME_INSTALL%\setup_viame.bat"
 
-python.exe "%VIAME_INSTALL%\configs\launch_search.py" ^
+python.exe "%VIAME_INSTALL%\configs\search.py" ^
   -qp pipelines\query_retrieval_and_iqr.pipe
