+++ examples/archive_summarization/launch_search_interface.sh
@@ -4,5 +4,5 @@ export VIAME_INSTALL="$(cd "$(dirname ${BASH_SOURCE[0]})" && pwd)/../.."
 
 source ${VIAME_INSTALL}/setup_viame.sh
 
-python ${VIAME_INSTALL}/configs/launch_search.py \
+python ${VIAME_INSTALL}/configs/search.py \
   -qp pipelines/query_retrieval_and_iqr.pipe
