+++ examples/registration_and_mosaicing/generate_mosaic_for_list.bat
@@ -16,4 +16,4 @@ dir /s/b/o "%INPUT_IMAGE_PATTERN%" > input_list.txt
 kwiver.exe runner "%VIAME_INSTALL%\configs\pipelines\register_using_homographies.pipe" ^
                   -s input:video_filename=input_list.txt
 
-python.exe %VIAME_INSTALL%\configs\create_mosaic.py first_mosaic.png output-homog.txt input_list.txt --step 1
+python.exe %VIAME_INSTALL%\configs\mosaic.py first_mosaic.png output-homog.txt input_list.txt --step 1
