+++ examples/registration_and_mosaicing/generate_mosaic_for_list.sh
@@ -16,4 +16,4 @@ ls ${INPUT_IMAGE_PATTERN} > input_list.txt
 viame ${VIAME_INSTALL}/configs/pipelines/register_using_homographies.pipe \
       -s input:video_filename=input_list.txt
 
-python ${VIAME_INSTALL}/configs/create_mosaic.py first_mosaic.png output-homog.txt input_list.txt --step 1
+python ${VIAME_INSTALL}/configs/mosaic.py first_mosaic.png output-homog.txt input_list.txt --step 1
