+++ examples/size_measurement/calibrate_cameras.bat
@@ -25,6 +25,6 @@ REM   -q SQUARE_SIZE  Width of calibration square in mm (default: 80)
 REM   -s FRAME_STEP   Process every Nth frame (default: 1)
 REM   -g              Show GUI with detection results
 
-python.exe "%VIAME_INSTALL%\configs\calibrate_cameras.py" -a -q 80 -o calibration_matrices.json %*
+python.exe "%VIAME_INSTALL%\configs\calibrate.py" -a -q 80 -o calibration_matrices.json %*
 
 pause
