+++ tools/configs.h
@@ -2,8 +2,8 @@
  * BSD 3-Clause License. See either the root top-level LICENSE file or  *
  * https://github.com/VIAME/VIAME/blob/main/LICENSE.txt for details.    */
 
-#ifndef VIAME_TOOLS_GET_CONFIGS_H
-#define VIAME_TOOLS_GET_CONFIGS_H
+#ifndef VIAME_TOOLS_CONFIGS_H
+#define VIAME_TOOLS_CONFIGS_H
 
 #include <vital/applets/kwiver_applet.h>
 
@@ -12,11 +12,11 @@
 namespace viame {
 namespace tools {
 
-class VIAME_TOOLS_APPLETS_EXPORT get_configs_applet
+class VIAME_TOOLS_APPLETS_EXPORT configs_applet
   : public kwiver::tools::kwiver_applet
 {
 public:
-  PLUGIN_INFO( "get-configs",
+  PLUGIN_INFO( "configs",
                "Extract pipeline and training parameters as JSON.\n\n"
                "Reads KWIVER pipeline (.pipe) or training configuration "
                "(.conf) files, either a single file or a whole directory, and "
@@ -30,4 +30,4 @@ public:
 } // namespace tools
 } // namespace viame
 
-#endif // VIAME_TOOLS_GET_CONFIGS_H
+#endif // VIAME_TOOLS_CONFIGS_H
