+++ tools/disparity.py
