+++ tools/inspect_file.py
