+++ tools/run_bulk.py
@@ -1513,21 +1513,21 @@ if __name__ == "__main__" :
 
   # Build out detection vs time plots for both detections and tracks
   if args.detection_plots:
-    import plot_detections
+    import plot
     log_info( lb1 + "Generating data plots for detections" )
     detection_plot_dir = os.path.join( args.output_directory, "detection_plots" )
     create_dir( detection_plot_dir, logging=False, recreate=True, prompt=False )
-    plot_detections.detection_plot( args.output_directory,
+    plot.detection_plot( args.output_directory,
       detection_plot_dir, args.objects.split( "," ), float( args.plot_threshold ),
       float( args.frame_rate ), int( args.smooth ),
       ext = detection_ext, top_category_only = False )
 
   if args.track_plots:
-    import plot_detections
+    import plot
     log_info( lb1 + "Generating data plots for tracks" )
     track_plot_dir = os.path.join( args.output_directory, "track_plots" )
     create_dir( track_plot_dir, logging=False, recreate=True, prompt=False )
-    plot_detections.detection_plot( args.output_directory,
+    plot.detection_plot( args.output_directory,
       track_plot_dir, args.objects.split( "," ), float( args.plot_threshold ),
       float( args.frame_rate ), int( args.smooth ),
       ext = track_ext, top_category_only = True )
